update clase 14

+++ Clases/C14S-DAO/src/com/company/dao/IDao.java
@@ -6,6 +6,6 @@ public interface IDao <T>{
 
     public T guardar (T t);
 
-    public T buscar (Long id);
+    public T buscar (Integer id);
 
 }
